requirements.txt
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-magic==0.4.27